+++ tests/todo.test.js
@@ -1,0 +1,10 @@
+'use strict';
+
+const ToDo = require('../src/components/todo/todo.js');
+
+describe('Testing ToDo Component, and dependents', () => {
+  it('Should', () => {
+    ToDo._addItem();
+    expect().toBe();
+  });
+});
